+++ packages/shared/src/index.ts
@@ -3,6 +3,7 @@ export * from './server-os';
 export * from './dto';
 export * from './utils';
 export * from './error-codes';
+export * from './api-error-contract';
 export * from './i18n-locale';
 export * from './slug-rules';
 export * from './git-provider';